res changed csVirtualClock to support suspend and return a current   time of 0 (instead ~0) in case it was not advanced yet (the latter   valued confused at least spr3d. 0 seems much saner in the context as   well.) res converted spr3d, its loaders and the sprcal3d loader to new SCF. res removed some methods from iSprite3DState that can be found in   iMeshObject as well.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@22971 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ libs/csgfx/renderbuffer.cpp
@@ -68,7 +68,7 @@ void* csRenderBuffer::Lock (csRenderBufferLockType lockType)
   {
     if ((lockType > CS_BUF_LOCK_READ) 
       || (props.lastLock > CS_BUF_LOCK_READ)
-      || (lockType != props.lastLock))
+      || ((uint)lockType != props.lastLock))
     {
 #if defined(CS_DEBUG) && defined(DEBUG_LOCKING)
       if (lockStack)
